completed getters
git-svn-id: svn+ssh://svn.savannah.nongnu.org/gap/trunk@3382 d25b4930-3292-e311-b074-0010c6de5da8

+++ system-apps/StepSync/FileObject.h
@@ -40,7 +40,7 @@
 - (void)setAbsolutePath:(NSString *)path;
 - (NSString *)relativePath;
 - (void)setRelativePath:(NSString *)path;
-- (unsigned long)size;
+- (unsigned long long)size;
 - (void)setSize:(unsigned long)size;
 - (NSDate *)createdDate;
 - (void)setCreatedDate:(NSDate *)date;
